Generalize script loader and add function for loading ROOT file geometry
