+++ src/app/sections/cms/cms.component.ts
@@ -4,7 +4,7 @@ import { PresetView } from '../../services/extras/preset-view.model';
 import { EventdisplayService } from '../../services/eventdisplay.service';
 import { HttpClient } from '@angular/common/http';
 import { CMSLoader } from '../../services/loaders/cms-loader';
-import { loadJSRootScripts } from '../../services/loaders/jsroot-loader';
+import { ScriptLoader } from '../../services/loaders/script-loader';
 
 @Component({
   selector: 'app-cms',
@@ -31,7 +31,7 @@ export class CMSComponent implements OnInit {
 
     this.eventDisplay.init(configuration);
 
-    loadJSRootScripts((JSROOT) => {
+    ScriptLoader.loadJSRootScripts((JSROOT) => {
       this.eventDisplay.loadRootJSONGeometry(JSROOT, 'https://root.cern/js/files/geom/cms.json.gz', 'CMS Detector', 10);
     });
